Move component ConfirmedCasesAttributesCard from 'components/cards' to 'components/index/CardsReference'
