+++ components/index/CardsReference/ConfirmedCasesAttributesCard.vue
